+++ lib/chat_models/chat_ollama_ai.ex
@@ -468,6 +468,8 @@ defmodule LangChain.ChatModels.ChatOllamaAI do
           IO.inspect(response, label: "RAW REQ RESPONSE")
         end
 
+        Callbacks.fire(ollama_ai.callbacks, :on_llm_response_headers, [response.headers])
+
         case do_process_response(ollama_ai, data) do
           {:error, reason} ->
             {:error, reason}
@@ -527,7 +529,9 @@ defmodule LangChain.ChatModels.ChatOllamaAI do
         )
     )
     |> case do
-      {:ok, %Req.Response{body: data}} ->
+      {:ok, %Req.Response{body: data} = response} ->
+        Callbacks.fire(ollama_ai.callbacks, :on_llm_response_headers, [response.headers])
+
         data
 
       {:error, %LangChainError{} = error} ->